fix(ci): unblock mattermost policy contracts on windows
