+++ extensions/mattermost/src/mattermost/monitor-auth.ts
@@ -1,73 +1,23 @@
-import { normalizeLowercaseStringOrEmpty } from "openclaw/plugin-sdk/text-runtime";
 import type { ResolvedMattermostAccount } from "./accounts.js";
 import type { MattermostChannel } from "./client.js";
 import type { OpenClawConfig } from "./runtime-api.js";
 import {
   evaluateSenderGroupAccessForPolicy,
   isDangerousNameMatchingEnabled,
-  resolveAllowlistMatchSimple,
   resolveControlCommandGate,
-  resolveEffectiveAllowFromLists,
 } from "./runtime-api.js";
+import {
+  isMattermostSenderAllowed,
+  normalizeMattermostAllowList,
+  resolveMattermostEffectiveAllowFromLists,
+} from "./policy-shared.js";
 
-export function normalizeMattermostAllowEntry(entry: string): string {
-  const trimmed = entry.trim();
-  if (!trimmed) {
-    return "";
-  }
-  if (trimmed === "*") {
-    return "*";
-  }
-  return trimmed
-    .replace(/^(mattermost|user):/i, "")
-    .replace(/^@/, "")
-    .trim()
-    ? normalizeLowercaseStringOrEmpty(trimmed.replace(/^(mattermost|user):/i, "").replace(/^@/, ""))
-    : "";
-}
-
-export function normalizeMattermostAllowList(entries: Array<string | number>): string[] {
-  const normalized = entries
-    .map((entry) => normalizeMattermostAllowEntry(String(entry)))
-    .filter(Boolean);
-  return Array.from(new Set(normalized));
-}
-
-export function resolveMattermostEffectiveAllowFromLists(params: {
-  allowFrom?: Array<string | number> | null;
-  groupAllowFrom?: Array<string | number> | null;
-  storeAllowFrom?: Array<string | number> | null;
-  dmPolicy?: string | null;
-}): {
-  effectiveAllowFrom: string[];
-  effectiveGroupAllowFrom: string[];
-} {
-  return resolveEffectiveAllowFromLists({
-    allowFrom: normalizeMattermostAllowList(params.allowFrom ?? []),
-    groupAllowFrom: normalizeMattermostAllowList(params.groupAllowFrom ?? []),
-    storeAllowFrom: normalizeMattermostAllowList(params.storeAllowFrom ?? []),
-    dmPolicy: params.dmPolicy,
-  });
-}
-
-export function isMattermostSenderAllowed(params: {
-  senderId: string;
-  senderName?: string;
-  allowFrom: string[];
-  allowNameMatching?: boolean;
-}): boolean {
-  const allowFrom = normalizeMattermostAllowList(params.allowFrom);
-  if (allowFrom.length === 0) {
-    return false;
-  }
-  const match = resolveAllowlistMatchSimple({
-    allowFrom,
-    senderId: normalizeMattermostAllowEntry(params.senderId),
-    senderName: params.senderName ? normalizeMattermostAllowEntry(params.senderName) : undefined,
-    allowNameMatching: params.allowNameMatching,
-  });
-  return match.allowed;
-}
+export {
+  isMattermostSenderAllowed,
+  normalizeMattermostAllowEntry,
+  normalizeMattermostAllowList,
+  resolveMattermostEffectiveAllowFromLists,
+} from "./policy-shared.js";
 
 function mapMattermostChannelKind(channelType?: string | null): "direct" | "group" | "channel" {
   const normalized = channelType?.trim().toUpperCase();
